Modify key to server mode so it performs multiple mappings per execution
The key to server mode (option -k), only ran 1 mapping per execution. If we
need to know the server for many keys, this can be too slow. This commit
changes the key to server mode so it reads keys from stdin instead of reading
just the 1 passed as an argument to the program.